Use 16-bit indices all the way from Cal3D to the renderer, which eliminates a bunch of copies and caches.  This change also gives us the ability to optimize the CalSubMesh::Vertex structure for cache utilization.
git-svn-id: svn+ssh://svn.corp.imvu.com/home/svnrepos/trunk/stable/Source/cal3d@49230 07c76cb3-cb09-0410-85de-c24d39f1912e

+++ src/cal3d/global.h
@@ -12,7 +12,7 @@
 
 #include "cal3d/platform.h"
 
-typedef int CalIndex; 
+typedef unsigned short CalIndex;
 
 namespace Cal {
